Break out generatePagesForCause
